requirements.txt
absl-py==2.2.1
addict==2.4.0
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.1
albucore==0.0.17
albumentations==1.4.18
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
args==0.1.0
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.3.0
backcall==0.2.0
blinker==1.8.2
Brotli @ file:///croot/brotli-split_1714483155106/work
cachetools==5.5.2
certifi @ file:///home/conda/feedstock_root/build_artifacts/certifi_1725278078093/work/certifi
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
click==8.1.8
clint==0.5.1
colorama==0.4.6
comm==0.2.2
ConfigArgParse==1.7
contourpy==1.1.1
coverage==7.6.1
cupy-cuda113==10.6.0
cycler==0.12.1
dash==3.0.1
dataclasses-json==0.6.7
decorator==5.2.1
Deprecated==1.2.18
docker-pycreds==0.4.0
entrypoints==0.4
eval_type_backport==0.2.2
executing==2.2.0
fastjsonschema==2.21.1
fastrlock==0.8.3
Flask==3.0.3
fonttools==4.56.0
freetype-py==2.5.1
frozenlist==1.5.0
fsspec==2025.3.0
future==1.0.0
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.38.0
google-auth-oauthlib==1.0.0
grpcio==1.70.0
hydra-core==1.3.2
icecream==2.1.4
idna==3.10
imageio==2.35.1
importlib_metadata==8.5.0
importlib_resources==6.4.5
index_max==0.0.0
intel-cmplr-lib-ur==2025.1.0
intel-openmp==2025.1.0
ipycanvas==0.13.3
ipyevents==2.0.2
ipython==8.12.3
ipywidgets==8.1.5
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==7.4.9
jupyter_core==5.7.2
jupyterlab_widgets==3.0.13
kiwisolver==1.4.7
lazy_loader==0.4
lietorch==0.4
lightning-utilities==0.11.9
llvmlite==0.39.1
mamba==0.11.3
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.22.0
matplotlib==3.7.5
matplotlib-inline==0.1.7
mdurl==0.1.2
MinkowskiEngine==0.5.4
mkl==2025.1.0
mkl-fft @ file:///croot/mkl_fft_1695058164594/work
mkl-random @ file:///croot/mkl_random_1695059800811/work
mkl-service==2.4.0
mmcv==1.3.15
mmengine==0.10.7
mpmath==1.3.0
multidict==6.1.0
mypy_extensions==1.1.0
narwhals==1.32.0
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.1
ninja==1.11.1.4
numba==0.56.4
numpy==1.24.3
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.2
omegaconf==2.3.0
open3d==0.19.0
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
OpenEXR==3.3.3
packaging==24.2
pandas==2.0.3
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow @ file:///croot/pillow_1721059439630/work
pkgutil_resolve_name==1.3.10
platformdirs==4.3.6
plotly==6.0.1
pointnet2==3.0.0
prompt_toolkit==3.0.50
propcache==0.2.0
protobuf==3.20.0
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybind11==3.0.0
pycocotools==2.0.7
pydantic==2.10.6
pydantic_core==2.27.2
pyglet==2.1.3
pygltflib==1.16.4
Pygments==2.19.1
PyOpenGL @ file:///home/mxh24/pyopengl
PyOpenGL-accelerate==3.1.9
pyparsing==3.1.4
pyquaternion==0.9.9
pyrender==0.1.45
PySocks @ file:///tmp/build/80754af9/pysocks_1605305779399/work
python-dateutil==2.9.0.post0
pytorch-lightning==2.0.0
pytz==2025.2
PyWavelets==1.4.1
PyYAML==6.0.2
pyzmq==24.0.1
referencing==0.35.1
requests @ file:///croot/requests_1721410876868/work
requests-oauthlib==2.0.0
retrying==1.3.4
rich==14.0.0
rpds-py==0.20.1
rsa==4.9
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
sentry-sdk==2.24.1
setproctitle==1.3.5
six==1.17.0
smmap==5.0.2
stack-data==0.6.3
sympy==1.13.3
tbb==2022.1.0
tcmlib==1.3.0
tensorboard==2.14.0
tensorboard-data-server==0.7.2
termcolor==2.4.0
threadpoolctl==3.5.0
tifffile==2023.7.10
tomli==2.2.1
torch @ file:///home/mxh24/torch-1.12.1%2Bcu113-cp38-cp38-linux_x86_64.whl#sha256=4adf483ac2d047534a7d023f0022bd8694d87627068ad6dddf186cb3273bbfa2
torch-scatter @ file:///home/mxh24/codes/torch_scatter-2.0.9-cp38-cp38-linux_x86_64.whl#sha256=f375384acd79963020be5643cda33792eab7aabae9083063a0e81bfc5cf0b6ea
torchaudio @ file:///home/mxh24/torchaudio-0.12.1%2Bcu113-cp38-cp38-linux_x86_64.whl#sha256=07324a67c069a986bb991c509c9442e8fce72326941e55fa4738d88cbb806047
torchmetrics==1.5.2
torchvision @ file:///home/mxh24/torchvision-0.13.1%2Bcu113-cp38-cp38-linux_x86_64.whl#sha256=899cac298d2a7cf6a8ca62d3ede2a7d3f50b86027f8be2d15639baf74da5d8f0
tornado==6.1
tqdm==4.67.1
traitlets==5.14.3
trimesh==4.6.5
typing-inspect==0.9.0
typing_extensions @ file:///croot/typing_extensions_1715268824938/work
tzdata==2025.2
umf==0.10.0
urllib3 @ file:///croot/urllib3_1727769808118/work
usd-core==22.5.post1
wandb==0.19.8
wcwidth==0.2.13
Werkzeug==3.0.6
widgetsnbextension==4.0.13
wrapt==1.17.2
yapf==0.43.0
yarl==1.15.2
zipp==3.20.2
zmq==0.0.0
